ComputerPlayer: Extend card action choice processor service for linked choices and use
This patch makes use of the newly added CardActionChoiceProcessorService
from the previous commit, in the ActionChoicePlayerService. This is used
for picking dwarves, cards, and actions (since all three are linked; no
point using a specific dwarf if you then forget, and decide to use a
card which doesn't care about the dwarf weapon level instead).