ComputerPlayer: Widen net cast by 'mark_invalid_action' over expedition rewards
This change makes the logic around which actions are marked as invalid
more intelligent, so more of the decision tree can be discarded earlier.
Specifically, actions that are the same level and higher are discarded,
since if the weapon isn't strong enough to get rewards at level x, it
won't be able to get rewards at level y>=x.