Logging: Replace string error with class when unable to afford to perform action
In four places, the player must purchase an item (placing a tile/twin tile/
stable, or purchasing a weapon). Prior to these changes, the error returned
when the item could not be afforded was just a simple string. This didn't
play very nicely with indenting lines, or allow for potentially colouring
the text, and so we're now taking advantage of python's duck typing to pack
a class into the ResultLookup errors field.

This new class has __str__ and __format__ defined, and returns mostly the
same error as before at this point, though the formatting is slightly
nicer, because we properly format the resources dictionaries. The format
method also optionally takes a number of spaces to pad the start of the
additional lines with.