KeyboardPlayer: Only print cards which can be immitated when user requests
This is a weird hack of a solution, which I'm not super happy with, but
it'll suffice for now. The first time through the loop the question is
skippable, and using the "q" does skip it, after printing additional
information. The next time through the loop, the question is no longer
skippable, and there's no "q" option/keybinding.

Might be better to use expand? Or change the UI entirely? Show the cards
in an actual list, and allow it to be skippable?