Logging+Runner: Include dwarf index when printing player information at start of turn
This would also benefit from printing "how many dwarves total".

Also, this should probably be moved into the keyboard player service? -- No,
probably not, because you want to say when the computer is taking a turn.
But the logging should depend on what render mode the consumer has defined
(this is currently just being completely ignored, problem for when we turn
training on)