Cards+Runner: Reduce number of ways of combining actions
Removes the AndOptionally combination enum in favour of OrAndThen, since
the former was only used in the Housework Card.

Additionally, change the logging card to remove the option where you only
go on an expedition, and do not take the items on the card, since that's
never the correct call.