KeyboardPlayer: Port query for position to place {tile,twin tile} to InquirerPy's new helper
This now uses the 'number' option, which auto validates that the player
has not tried to input a non-number.

Also supports spinning, though this currently goes through all of tile
positions between the minimum and maximum, the majority of which are
probably not valid placements. Would be nice if the arrow keys instead
changed the input to the next valid placement instead, which should be
possible by defining a custom keybinding.