KeyboardPlayer: Add keybinding in "tile to purchase" keyboard query to print tile cost
Struggled for _way_ too long trying to get this to work with itertools
stuff, and zips and cycles, and then said "screw this" and did it in 30
seconds using a loop. yay.

Anyway, this is much needed functionality (otherwise you couldn't see how
much a tile cost). Also, quantities of resources are yellow now! Which
makes them a lot easier to see.

Also included a long-instruction, to tell the user that they can do this.

Probably worth also adding one to get the current resources, which should
be accessible in a lot of places.

Briefly tried making the resource name green, but that didn't end up
looking very good. Might be wroth revisiting.

This should be extracted into its own method, so that it can be used
anywhere that we want to convert resource costs into a printable string.