KeyboardPlayer: Stop tile position spinning from going through invalid locations
The default 'number' prompt supports using arrow keys to 'spin' between
different number inputs. Though useful, the majority of positions were
not valid locations to place the tile, and so the user had to spend a
long time spinning between different inputs to pick their desired
placement.

This patch overrides the base behaviour (hackily) so that now, when the
up/down arrow keys are pressed, we jump to the next/previous valid
location, which is a lot more intuitive to use! ^^

Additionally, this patch removes the ability to enter numbers into the
number prompt, because that wasn't working well -- any value just set
the value to the maximum.

Potential further work: it would be nice to highlight the chosen
location in the minimap. Would probably require changing that to be
part of the buffer? Which sounds tricky.