KeyboardPlayer: Pretty Print cost of tile when picking effects to use
This patch brings the "effects to use" prompt up to date with recent
changes, adding the ability to check the current player resources, and
pretty printing the cost of the tile. This works in a similar way to
the "imitate a card" prompt, where it skips the prompt when the player
requests the card, and then loops back to the start of the prompt.

Still doesn't work, though, if the player says that they actually want
to use an effect, since we're not currently fetching the effects. But
that's how it was before, anyway, so no loss there.