Runner: Do not explode if all action outcomes end up with negative score
Prior to a recent change, we were relying on the permutation ordering
operation returning a ResultLookup with not-none value. That assumption
didn't hold for the point permutation ordering service, because it used
to only consider permutations as valid if they ended up with positive
score, which didn't hold because at the beginning of the game the score
is often negative (while large swathes of the board have no tiles).

After this change, we've dropped the assumption that scores are
positive, and so are always returning a valid ordering object that we
can iterate through.