Runner: Fix buying a tile resulting in it being unavailable to purchase
None of the "place a tile" actions are pure. Because they need to mark the
tile as purchased, when invoked they mutate the turn_descriptor_lookup,
removing the just-purchased tile from the list of tiles available to buy.
The problem comes when this process is run as part of the action ordering
step; the first time the actions are attempted to be laid in order, the
tile is purchased. Subsequent times through, the tile is unable to buy,
since it was purchased the first time.

Made the exhaustive action ordering service (currently the only ordering
service) restore the state of the turn_descriptor_lookup to as it was
before each ordering is trialed, which solves the issue and allows the
player to purchase tiles.