Action+Players: Make placing fences its own action
The combinatorics for the placing of actions was broken, and was also
resulting in an unwieldly number of entiries when trying to perform the
sheep/donkey farming actions.

It's now been switched into a conditional action, which allows the prompt
to be a lot nicer. Not currently implemented for the action choice player,
but the base action has been added so that things still compile.