Cards: Make card action combinations match reality
Turns out that, after much research (reading the appendix), "And Then
Or" does not mean "a then b, or just b". The conditional service has
been updated to match this.

However, it does still make sense to have a "a then b, or just b" case,
to reduce the combinatoric complexity -- for example, when taking
accumulated items, and the refill is only ever better than if not taking
resources always. This case is now represented in OrAndThenStrict.

Formatting to come in the next commit.