Actions(+Style): Fix expeditions failing when picking weapon upgrade reward
Precognitive todos ahould be listened to. We were treating no
successfully upgraded weapons as a fail result, and that was causing
the entire exhaustive search to fail. Not sure why this wasn't
printing any errors to the user though -- possible something in the
deduplication code?

This patch also changes the code style to return early.