KeyboardPlayer: Stop colour printing of take accumulated items from throwing
The new colour printing just extends the text to be printed with the
formatted output. Because of python's horrible duck typing, if the
formatted output is just a string, that adds each char individually,
and since those arent styled tuples, the colour printer fails.

This patch just fixes it for the one case I saw, but it's possible
there are other issues (whenever not using a multiconditional). We
should probably force the output to be a colour printable type.