Actions: Manually enforce that animal storage is checked only once
The CheckAnimalStorageAction used to return a Uniqueness constraint, which
was never implemented. This was causing the weekly market action to fail
when trying to purchase multiple animals. The constraint has now been dropped
(and the class removed) in favour of manually checking in the turn execution
loop. Hopefully this will play nicely with the constraint that this check
must be performed after the animals are received.