Runner+Logging: Add more context to logs when ordering permutations
Now prints the total number of permutations considered, so that we can
see how many were run after the last "valid" permutation (we already get
logging of permutations which pass the constraints).

Renames "permutation"->"action" when going through permutation and saying
if an action succeeds or fails.

Adds a str representation to the check animal storage action.