Actions: Reduce severity of invoking an expedition with no chosen rewards
This happens if you just spam through the expedition reward choices
without picking any, which has been happening a lot when I've been
testing. No reason that it should cause the entire game to crash.

However, this is a nonvalid use case which should really be handled
by chosing the card actions which do not involve going on an
expedition, so it still fails the permutation check. Not decided if
this is totally sensible though.