KeyboardPlayer: Try to wrap text when displaying potential action choices
Thank you SIBR for introducing me to the textwrap functionality.

Unfortunately, we regret to inform you that the textwrap functionality
does not work with pretty formatted action strings, since those are
made up of many parts. So this doesnt really do anything in any of
the test cases so far.

But it could!