KeyboardPlayer+Action: Add more context to weekly market prompt
Changed the signature of the get_player_choice_market_items_to_purhcase
method, so that it now includes the purchasable items, and the cost to
purchase them.

Add a keybinding to print the current total amount of coins the player
must pay, and also an automatic warning if the amount is more than the
amount of coins the player is expected to have after receiving their
pocket money for going to market.

This would be a good candidate for splitting onto multiple lines, when
that method is fixed.