Actions: Don't attempt to build the same tile forever when invoking an action multiple times
The place a single tile action was using the specific tile generation method
as a way of determining whether or not to get the player choice of which
tile to build. Since this wasn't being nulled out (because it doesn't make
sense to null out the constructor argument), we need to check some other
value, which this patch has now added.