KeyboardPlayer: Allow player to pick conversion effects
Add a multistep process for converting resources into food (just food
at the moment, others to come. This isn't limited to only the starting
food though, it supports anything which ends up at food). This is done
by:
 1. asking the player if they would like to convert food. Defaults to
    no, for ease of speedy clicking. Probably makes sense to keep it this
    way; most turns you're not trying to get more food.
 2. presenting the user with a list of resources that could be converted
    into food. This works okay, but the current implementation just shows
    the user the cost of whatever conversions they have. This leads to
    duplicate entries in the combobox prompt -- for coins and for donkeys
    -- which should be cleaned up.
 3. for each of the choices the player made, get the number of times that
    they'd like to invoke the input. Like the above prompt, this doesn't
    work especially well with the effects where there are multiple choices
    (particularly coins, where it might be possible to convert fewer coins,
    if you picked one of the expensive options).
All of these steps are supportive of the player backing out; if they pick
no conversions after saying they wanted to, or they say that they don't
actually want to pay any of the costs, the conversions are simply just not
included in the output.

This has highlighted a bug with the conversion effects -- when converting
coins, and chosing the "two coins" option, the ConvertSingleAction, seems
to get stuck trying to find a conversion effect. Investigation to come.