Logging: Improve logging for conversion actions
Trying to just run the __format__ method for conversion as part of string
interpolation was failing because the format_spec was "", and that is
apparently not isspace. Switched to just comparing against the truthy
value instead now, because that apparently does what is expected.

Might be better to use direct comparison against the empty string though,
and this might need fixing in the other __format__ overrides.