Action: Give resources when placing twin tiles over resources
Added a new class extending BaseTile; TransparentTile, which is only
here to give a nice constructor for thing which mimic a tile and have a
single receive effect.

We're then just manually checking for these at the two locations when
placing the tile, and converting and resources into receive actions.

Tested with overhanging tiles, with food, and with boars (individually).
All seemed to work, and receiving a boar successfully added a "check
animal storage" action and corresponding constraint.

Currently only implemented for twin tile placement; single tile placement
to come.