Action: Give resources when placing single tiles over resources
Also fixes the bug when trying to placed non-unique tile, and unable to
remove from turn_descriptor.tiles, which was preventing pastures from
being placed (among other things, presumably).

Still untested (thought that it could be tested using the sheep farming
action, since that allows for placing single pastures, but this only
allows for things to be built over already-existing infrastructure).
Could be tested with the expedition, but can't currently work out a way
of getting the dwarf weapon up to level 9 within 3 turns.