Runner: Ask for desired conversions after asking for action specific choices
We get the player choice for actions in the order that they are added, and so
by inserting the convert action after the card actions, they're processed in
a more natural order; usually players only want to convert resources at the
end of the turn, after they've already chosen the actions to perform.

Note: this reordering is not perfect. If the player goes on an expedition,
and chooses an action which places a tile, they will be required to make a
second round of choices, which, in the current implementation, will come
after the conversion effects. Which is potentially confusing (where did my
choice go), and also sub-optimal (wait I wanted to spend the coins I'm going
to get from overhanging).