Optimisation+Runner: Speed up action ordering by storing factorials as variables
The permutation forge uses factorials in order to work out which value
should come next in the permutation. This patch precalculates the factorials
at the beginning of the method, and then reuses them throughout. This reduces
the number of calls to math.factorial to 36% of the previous value (from
3265921 down to 9072001).

Also, do the same for len; down to 20% of previous value, (3266206 -> 15604126).

This results in a net speed increase to the turn execution code of 40%!