Optimisation+Runner: Cache permutation index calculation results
For each element in a list to-be permuted, we calculate an index and
a remaining index, which additionally requires calculating the factorial.
This patch extracts that calculation, replaces the previous modulus
and division/floor methods with a single call to divmod, and now caches
the result of any inputs, using the functool.cache decorator.

This reduces the number of factorial calls from 3,265,921 to 2,903,041
without caching, and 409,113 with caching, as well as removing all
2,903,040 math.floor invocations and replacing them with 409112 divmod
invokes. Together, this brings about about an 11% speedup, from 12.4s
to 11.0s.