Optimisation+Runner: Allow invalid action permutations to be pruned early
Once we've resolved that the permutation doesn't pass due to issues in
some early actions, there's no point continuing to shuffle around the
remaining items in the permutation to see if it works -- since they're
downstream of whatever causing it to fail, they can't make it work.

This patch adds the functionality to allow that to be discarded. First,
the constraint class was modified to return at what point it fails, and
this data then used in the permutation forge, in order to increment the
current index to the next potentially valid permutation, skipping all
of the known-to-be bad permutations.

No idea if this works! It doesn't fail, at least. And it's a hell of a
lot faster -- down from 10s to 0.04s, and reducing the number of
permutations considered from 362,880 to 536. The output also appears to
show the sort of pattern I'd expect, though again, not yet done a
complete evaluation of the output. Additionally, the test case I've
been considering has no valid cases, and so isn't actually running any
of the actions to see if those failing also cause the pruning to occur.