Optimisation+Runner: Speed up constraint validation stage
When the precedes constraints are validated, they require the index of
the actions they cover, in order to determine which is first. The old
method of getting these indexes was to walk the list of permutations
until we found them, checking each time.

This was slow for multiple reasons; first, algorithmically; it's either
O(n) time to iterate through the permuted actions, or  O(n^2) when
considering that we need to search the permutation for the matching
actions for every constraint that has been applied.

Secondly, the way that we were checking that the action matched was to
do an equality comparison between the target action and the iteration
action, which meant that we were doing this a lot. This showed up in the
profile output as a large amount of time spent in specifically
isinstance, since this is usually the first thing we do -- check that
the two actions we're comparing are even of the same type.

The new implementation, introduced in this patch, hands the validation
of the constraints off to a new service dedicated specifically for that
task. The first optimisation in this service is to pre-cache the action
indicies in a dictionary, so that we can do an O(1) lookup of the index
instead of searching all of the actions. It also means that we amortize
the cost of the equality comparison, in favour of computing the hash of
the target action a single time instead. This reduces the number of
operations from ~24,855,682 equality comparisons to ~4,014,971 hash
lookups, which in turn reduces the time to execute from 8.0s to ~1.2s.

Secondly, we collate constraints, bundling those for which the first
action which must precede are the same. Not got any firm numbers for
the benefit of this, but it should mean that we're nearly cutting the
number of index lookups (and so hash computations) in half, by only
needing to get the index of the preceding action once, rather than per
constraint.

Result of testing: for some reason, after these changes, the number of
permutations considered went up from 38554 to 59904. Not sure why this
happened; more invesitgation to follow.

Edit, investigation results: There was no way of telling the difference
between multiple pay actions, and so we were just using the index of
the last (pruning)/first (bulk) action we put into the dictionary
(pruning)/got to (bulk). Fixed in a different commit.