Runner: Remove circular precedes constraint on harvest actions
Like all actions, the ResolveHarvestAction must come before any actions
that it spawns (HarvestField, FeedDwarves and BreedAnimals). However,
due to some overzealous constraining, we were enforcing that the
ResolveHarvestAction must come after _every_ other action. Including
the actions it spawned.

This constraint has now been relaxed; the ResolveHarvestAction only
requires that it is after the actions directly added by the card, not
child actions. Not sure if this is the correct call -- fewer constraints
mean we're considering more permutations, which isn't wrong (because if
we pass constraints then it's a valid permutation and should be
considered), but it also means that it's possible for, for example,
expedition rewards to happen after the dwarves have been fed.