Bugfix+Runner: Require that spawned actions happen before harvest
This forces resolve harvest to happen after all spawned children of
card actions. It also, potentially undesirably, forces the harvest to
happen after all of the spawned conversion actions. This seems to match
the rules but I'm not sure that it's completely sound.

Additionally ensures that there are not multiple identical constraints
that an action has to happen before the animal storage is checked.