Runner: Add flag to help with debugging permutation generation
I miss the good old days where none of the permutations actually worked.
That was fast! To get back to that; this patch introduces a flag to the
bulk ordering service that bails early when we pass the constraint
checks, in order to avoid the cost of the action invocation.