Optimisation+Runner: Fix memory leak in list permutation forge
The profile was being forcably shut down by linux after considering
about 290K permutations (error message simply "Killed"). Tracked this
down using tracemalloc to a memory leak in the pruning list permutation
forge; the cache that was being kept on _get_indicies was growing
without end. ...because we pass the index into this method, and the
index just grows as we continue through the permutations.

The cache does seem to still be useful though; switching to using the
lru_cache decorator, and checking the cache_info instrumentation, the
number of hits vs misses seems to be around 40%, which isn't bad. Going
to profile just to be sure.

Results from profile: Lol. Lmao.
The cache slows things down -- ran a test where the maximum permutation
index was capped at 450k; the cached version (with a cache size of 1024)
took 81.2s, and the non-cached version took 79.4s. Not much of a
difference, and we did spend more time in divmod/factorial, but that
wasn't a factor compared with the increased time spent in "generate list
permutation for index" (from 37.7s cumulative/28.4s direct when caching,
to 36.6s cumulative/24.5s direct without).