Optimisation+Logging: Make debugging of constraint validation more verbose
This patch extracts out a lot of the functionality in the constraint
validation code into individual methods, so that it shows up more
nicely in the profile (wish there was a way of doing this using the
profile library directly). This might add a pretty big overhead --
the profile time jumped up by 8s.

Seems like it's mostly in getting the action hash, which isn't
unexpected. Improving memory allocation in the list permutation
forge might be a better route to take, as that's the other big chunk
of time.

Additionally:
 * Include invoking actions in ordering again (didn't actually do
anything, still seems like they all fail permutation ordering).

 * Reduce the number of lines printed in formatted profile (10% was
about the cutoff point where the number of calls started to drop
significantly, so we're likely out of the loop).