Optimisation+Runner: Attempt to improve performance of pruning list permutation forge
Added a new version fo the permutation forge which is:
 1. "allocation free" (still does some slight allocations every iteration,
    but not as many, by virtue of being...)
 2. non-recursive.
Additionally, the factorial that we use in the calculation are now cached
in a list to be used in every iteration.

Not yet settled on the method used to convert the index_in_remaining_list
into the index_in_complete_list, since updating the offsets still takes
around 25% of the time spent generating the list permutation for index.