Optimisation+Runner: Stop single loop constraint validator from doing equality checks
The old version of doing this turned out to not be any quicker than
the collating version, due to cost in:
  * dictionary comprehension,
  * list comprehension, and
  * equality comparison between the current and 2nd action in the
    constraint.
This new patch gets rid of all of these in favour of calculating if the
action is the first or second when caching it in the constraints
dictionary. We also get around the dictionary/list comprehension by
setting a flag on whether the first action has already been seen,
instead of removing it from the list. Still not strictly pure, because
we have to reset these flags afterwards, but finally quicker!

This speeds up the constraint validation from 17.3s to 11.1s, a 36%
improvement! Finally. It's so late. I'm so tired.