Optimisation+Meta: Add full stack profile
Now that I've _finally_ gotten the constraints to pass for the test
case we've been running, it's worthwhile to test parts of the code
other than the permutation forge or the constraint validator.

This new profile is a lot slower, because it's actually doing things,
which necessitated dropping the total permutation count. Additionally,
the profile is a lot more noisy now (because we're hitting more files),
so the profile formatting has been adjusted to match.