Optimisation+Runner: Don't clone transparent tiles when creating SimplePlayer
Most board states don't result in any changes to the transparent
unavailable tiles, so there's no reason to copy from the real player
tiles onto a cloned simple player repository.

This new check educes the number of transparent tile -> generic tile
clones that we do to ~12% of the previous value (from 1,052,931 to
127,628). This... doesn't actually have any effect, because the gain
we get is lost in extra processing time in the tile_container_default.
Somehow?? No idea how, we haven't changed the default...