Optimisation+Container: Don't allocate memory when counting tiles
As the comment says, python is a stupid language. For some reason all
of the help documentation says that using sum is the best api here?

This isn't a hot path, is only called ~once in the ordering process
when calculating the best permutation to go with, but this is. Neater?