Optimisation+Tile: Cache whether tile has effects
The list comprehension that runs when we call "get_effects_of_type",
checks if the tile has any effects before fetching them. This caused
it to be hit 16,903,776 times in the current profile, and take up 12s
of the ~49s spent in get_effects_of_type. By caching the value instead,
has_effects now takes 2s, and results in a corresponding speedup in
get_effects_of_tye (now down to ~38s).