Optimisation+Tile: Cache effects by type
We're getting about 50% cache hit rate after this change, which brings
a corresponding 15s speed up over the version which has the callback
but does nothing with it.

The cache is cleared if any of the tiles update -- could do something
cleverer, perhaps, since we have the tile type? and would only need to
update the effects which have changed? Would need to do some testing to
determine if we're actually clearing the cache unnecessarily, which I'm
doubtful of at the moment.