Effects: Make effects declare if they hold state
When cloning the player (and so the tiles), we currently do a deepish
copy of all of the tiles and tile effects. This takes a long time, and
for most tiles isnt required, since the tile is unchanged between turns.

This patch makes each tile aware of if it stores state, which should be
used in future work to make this process more efficient -- maybe it's
not necessary to reset the entire player if the tiles aren't changing.