Bugfix+Actions: Don't fail if user has no food to feed dwarves
Due to some bugs in the enriched runner, we're currently seeing cases
where the amount of food that we have available when feeding dwarves is
zero (it all gets converted into begging markers). Due to an overzealous
guard clause in the resource container, this was previously throwing,
but it's a valid case, and should be accounted for.

Fixed by no longer trying to take zero food (the key is removed), and
also not trying to take zero amount of any resource (in the bulk take
method).

Additionally adds more context to the exception thrown when trying to
take an invalid amount of resources