Bugfix+Runner: Enforce that valid permutations are readonly when caching
Whawha. This took me multiple days to debug.

The so-called "allocation-free" permutation forge is allocation free. It
does this by reusing the array that it spits out, and overwriting the
values used in the last iteration. This means that if you were to say,
store a reference to the permutation that is spat out, it would change
underneath you.

Which is exactly what happened. After validation, the permutation that
was considered valid was being cached in a list, so that we could, at a
later step, find the best order. And then the next permutation was called
and changed the permutation we just stored.