Runner(+KeyboardPlayer): Enrich possible locations to build single tile
Because we only place tiles at the end of the turn, the player might
know better than us which location is valid. In order to give them
control over placements that we cannot verify, it is helpful to also
give them an enriched "valid locations" list, which can account for
tiles which we know will _never_ be valid (in this current
implementation, this only supports locations that are outside/inside,
when wanting to place a tile inside/outside. However, this could be
extended in the future; most furnished caverns (all but two) can never
be placed on another furnished cavern). This patch adds that enrichment.

First, we define a new type, the ValidLocations, which wraps a list of
LocationValidity, a new (messy) enum. This is then returned from the
"get_available_locations_for_single" method (FIXME: change the other
"get_available_locations_for_x" methods).