KeyboardPlayer: Allow user to place tiles in invalid locations
This loops because it adds some nice whitespace in between the different
printed maps. Having to do a weird hack with the message when skipping
the output, because using the filter doesn't work in that case.

Ideally we'd print the prompt message first, and then the map, but I
don't know how to do that, so this is what we've got. It's mostly fine,
it works well enough for the time being.

Would be nice to have some colours... And to potentially not have the
divider between inside and outside extend into the unavailable section
when none of the valid locations are unavailable.

Not currently working; if you try and place a stable, this'll break,
because the stable locations haven't been switched over to the new valid
location interface. Might be worth just making a new method for that
prompt, since there's no concept of adjacency/requisites for that?