KeyboardPlayer+Runner: Fix twin tile placement
Forgot to switch this over when the base class changed. Not ideal that
we're having to construct the tile twin placement lookup ourselves, but
probably preferable to returning it from the iter method anyway.

Also, did you know; if you only define __getitem__, python will try and
fall back to using that when iterating the value? Because that caught me
out.