Runner: Revert removal of requisite override param when placing single
I think this was removed because I was hitting both "get valid
locations" methods in the same refactor. I hope so, because if not,
removing the requisite override for twin tile placement was also
invalid. It was invalid for single because of the weirdness to do with
ruby mine construction (you can either place a ruby mine on a deep
tunnel, in which case you get extra goodies, or place it on a shallow
tunnel, in which case no goodies for you.).