Player: Add method for reporting invalid action choice to player
Currently just a stub, this should be used to inform the player that the
action that they tried to take was not valid, and request the kind of
modifications they'd like to try in order to generate a valid action
choice.

The enum values are pretty optimistic, not sure if there's actually a
use case for that many choices, of if it would be a pain to implement.
Tried to be exhaustive rather than conservative -- job for future me to
determine if all of these get to stay.