Refactor+Runner: Overhaul turn execution logic to allow changing choices
The previous implementation of the turn execution service (and child
turn transfer service) simply returned early if any of the steps failed.
This was not especially compatible with the idea of letting the user
make other choices if their previous ones resulted in being unable to do
something; e.g., if we failed action ordering because the player choice
to go on an expedition with a dwarf who didn't have a weapon, there was
no clear way to allow the user to only pick the dwarf again.

A quick rewrite in rust later, and we have the solution; state machine.
This allows for defining specific state transitions that occur, and step
between them. It also supports looping until fully complete.

All of the current logic lives in one place, which is not ideal --
should probably be split into multiple files. Additionally, we have to
carry information through from one state to the next, which seems a bit
messy -- why not just provide the {player, turn descriptor, whether it's
the player's final turn} in the `TurnState.next()` method?

Finally, after this patch, we're still not finished -- the full logic
around when an action fails is yet to be implemented, and will throw.
Not really a step back -- the old solution was no better, it just took a
little longer to crash.

The turn transfer service has been left for the time being, because it
might provide logic that could be useful when telling the player about
invalid actions.