Profile: Decorate heavily trafficked methods with line profiling
Added the requirement for package line_profiler. This is being used in
the methods that showed up as highly profiled; ordering actions,
checking if a permutation passed a constraint. Hopefully should be able
to point us in the direction of more useful fixes. (Has already found a
bug in an early return to avoid logging!)