ComputerPlayer: Allow action choice player to be constructed
This was left in a State. The aim of this patch is to make it slightly
less of a state.

All currently implemented action choice player methods have been made to
check that the action choice is set before attempting to read the action
choice. Currently left unimplemented because that's a lot of work and I
haven't thought about it in 2 years and fixing runtime exceptions was
easier. Any unimplemented methods have had their implementation
standardised, to just throw a 'NotImplmentedException'.

Resource to Sow processor service had an import fixed.

Tile purchase effect processor service was made to compile, but has no
logic and should probably have just been removed instead.

Nothing of consequence changed in the card processor, just changed from
using Typing.List to using list.