Runner: Allow game to be played with only bot players
The new 'bot_main' class allows for running the env with a fixed number
of 'action choice' players.

NB: This was done in a way that fit best with the current
implementation. However, that's not ideal -- since we only provide the
player default to the env, it's tricky to say e.g.,
 Player 1. Human
 Player 2. Computer
 Player 3. Human
 Player 4. Human
which would be made easier by extracting the player creation logic from
the gym. Still, this 'just works' so is what I went with for the time
being (long way off keyboard players playing against computers anyway).